% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dacomp_check_reference_set.R
\name{dacomp.check_reference_set_is_valid.k_groups}
\alias{dacomp.check_reference_set_is_valid.k_groups}
\title{Test reference set of taxa for validity, by multivariate K-Sample testing}
\usage{
dacomp.check_reference_set_is_valid.k_groups(X_ref, Y, nr.perm = 10^4,
  verbose = F)
}
\arguments{
\item{X_ref}{Matrix of counts, containing the reference taxa alone. Rows represent samples, columns represent taxa.}

\item{Y}{Categorical variable of the measure phenotype, the group labeling}

\item{nr.perm}{Number of permutations for computing P-values}

\item{verbose}{Should messages be printed to console, indicating which test is being computed.}
}
\value{
An object of type "dacomp.reference.validation.result.object", is a list with the follows fields:
\itemize{
\item{p.value.HHG.L2}{ - P-value for the HHG K sample test, with the L2 distance metric}
\item{p.value.HHG.L1}{ - P-value for the HHG K sample test, with the L1 distance metric}
\item{p.value.HHG.BC}{ - P-value for the HHG K sample test, with the Bray-Curtis dissimilarity metric}
\item{p.value.energy.L2}{ - P-value for the disco test, with the L2 distance metric}
\item{p.value.energy.L1}{ - P-value for the disco test, with the L1 distance metric}
\item{p.value.energy.BC}{ - P-value for the disco test, with the Bray-Curtis dissimilarity metric}
\item{p.value.permanova_L2}{ - P-value for the permanova test, with the L2 distance metric}
\item{p.value.permanova_L1}{ - P-value for the permanova test, with the L1 distance metric}
\item{p.value.permanova_BC}{ - P-value for the permanova test, with the Bray-Curtis dissimilarity metric}
}
}
\description{
The function carries out the Reference Validation Procedure (RVP), described in Brill et. al. (2019), appendix A. The test verifies no differentially abundant taxa have entered the reference set. The input to the function includes a counts matrix of taxa selected as a reference set (rows represent samples), and a vector Y of groups labels. The function first rarefies all samples to constant depth, then tests for multivariate equality of distributions across the differet study groups.
}
\details{
Rarefying samples to equal depth is mandatory for testing, see Brill et. al. (2019) Appendix A for a complete discussion of the test and assumptions. The multivariate K-sample tests performed include the \code{hhg.test.k.sample} (Heller et. al. 2019) from package \code{HHG}, The \code{disco} test (Székely et. al. 2007 ) from package \code{energy} and the Permanova test (Anderson et, al. 2001), available through the function \code{adonis} in the \code{vegan} package.
All three tests require a distance (or dissimilarity) metric between samples. All three tests are evaluated with the L1,L2 and Bray-Curtis dissimilarity metrics.
This function supports categorical phenotypes. For continous and multivariate phenotypes, a different procedure is required, making use of multivariate tests of independence such as \code{hhg.test} (from packge \code{HHG}), and \code{dcov.test} (from package \code{energy}).
}
\examples{
#' #' \dontrun{

library(dacomp)

set.seed(1)

data = dacomp.generate_example_dataset.two_sample(m1 = 100, 
       n_X = 50, n_Y = 50,
       signal_strength_as_change_in_microbial_load = 0.1)

#select references: (may take a minute)

result.selected.references = dacomp.select_references(X = data$counts,
                                 median_SD_threshold = 0.6, #APPLICATION SPECIFIC
                                 verbose = T)

length(result.selected.references$selected_references)

#plot the reference selection scores (can be used to better set the median SD threshold...)
dacomp.plot_reference_scores(result.selected.references)

result.ref.validity = dacomp.check_reference_set_is_valid.k_groups(
                        X_ref = data$counts[,result.selected.references$selected_references],
                        Y = data$group_labels,
                        nr.perm = 10000,
                        verbose = T)
result.ref.validity

} 
}
