% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dacomp_generate_example_data.R
\name{dacomp.generate_example_dataset}
\alias{dacomp.generate_example_dataset}
\title{Title}
\usage{
dacomp.generate_example_dataset(n_X = 30, n_Y = 30, m1 = 30,
  signal_strength_as_change_in_microbial_load = 0.1)
}
\arguments{
\item{n_X}{}

\item{n_Y}{}

\item{m1}{}

\item{signal_strength_as_change_in_microbial_load}{}
}
\value{
a list
\itemize{
\item{counts}{}
\item{group_labels}{}
\item{select_diff_abundant}{}
}
}
\description{
Title
}
\examples{
\dontrun{
library(dacomp)

set.seed(1)
data = dacomp.generate_example_dataset(m1 = 100,
       n_X = 50,
       n_Y = 50,
       signal_strength_as_change_in_microbial_load = 0.1)



} 
}
