% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dacomp_generate_example_data_continuous.R
\name{dacomp.generate_example_dataset_continuous}
\alias{dacomp.generate_example_dataset_continuous}
\title{Title}
\usage{
dacomp.generate_example_dataset_continuous(n, m1 = 30,
  signal_strength_as_change_in_microbial_load = 0.1)
}
\arguments{
\item{m1}{}

\item{signal_strength_as_change_in_microbial_load}{}
}
\description{
Title
}
\examples{
\dontrun{
library(dacomp)

set.seed(1)
data = dacomp.generate_example_dataset_continuous(n = 100,
m1 = 30,signal_strength_as_change_in_microbial_load = 0.1)



} 
}
