% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dacomp_generate_example_data.R
\name{dacomp.generate_example_dataset_multivariate_example}
\alias{dacomp.generate_example_dataset_multivariate_example}
\title{Title}
\usage{
dacomp.generate_example_dataset_multivariate_example(n, m1 = 30,
  signal_strength_as_change_in_microbial_load = 0.1)
}
\arguments{
\item{n}{}

\item{m1}{}

\item{signal_strength_as_change_in_microbial_load}{}
}
\value{
a list
\itemize{
\item{counts}{}
\item{covariate}{}
\item{select_diff_abundant}{}
}
}
\description{
Title
}
