% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceScore_Plot.R
\name{dacomp.plot_reference_scores}
\alias{dacomp.plot_reference_scores}
\title{Plot histogram of median SD statistics from reference selection object}
\usage{
dacomp.plot_reference_scores(ref_select,
  label = "Histogram of medianSD statistic, across taxa",
  quantiles_to_plot = c(0.5, 0.7, 0.9), breaks_param = 30)
}
\arguments{
\item{ref_select}{Computation result of \code{dacomp.select_references}}

\item{label}{Histogram label.}

\item{quantiles_to_plot}{Percentiles marked by vertical lines. Default value is \code{c(0.5,0.7,0.9)}.}

\item{breaks_param}{Passed as argument with the same name to the \code{hist} function.}
}
\description{
Plot histogram of median SD statistics from reference selection object
}
\examples{
\dontrun{
library(dacomp)

set.seed(1)

data = dacomp.generate_example_dataset.two_sample(m1 = 100,
       n_X = 50,
       n_Y = 50,
       signal_strength_as_change_in_microbial_load = 0.1)

#select references: (may take a minute)
result.selected.references = dacomp.select_references(X = data$counts,
                                                     median_SD_threshold = 0.6, #APPLICATION SPECIFIC
                                                     verbose = T)

length(result.selected.references$selected_references)

#plot the reference selection scores (can also be used to better set the median SD threshold)
dacomp.plot_reference_scores(result.selected.references)

} 
}
