% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Functions.R
\name{print.dacomp.reference.selection.object}
\alias{print.dacomp.reference.selection.object}
\title{Title}
\usage{
\method{print}{dacomp.reference.selection.object}(obj)
}
\arguments{
\item{obj}{}
}
\description{
Title
}
