% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Functions.R
\name{print.dacomp.result.object}
\alias{print.dacomp.result.object}
\title{Print the results contained in an object from type dacomp.result.object}
\usage{
\method{print}{dacomp.result.object}(obj)
}
\arguments{
\item{obj}{Object returned from dacomp.test(...)}
}
\description{
Print the results contained in an object from type dacomp.result.object
}
