% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcomp_generate_example_data.R
\name{wcomp.generate_example_dataset}
\alias{wcomp.generate_example_dataset}
\title{Title}
\usage{
wcomp.generate_example_dataset(n_X = 30, n_Y = 30, m1 = 30,
  signal_strength_as_change_in_microbial_load = 0.1)
}
\arguments{
\item{n_X}{}

\item{n_Y}{}

\item{m1}{}

\item{signal_strength_as_change_in_microbial_load}{}
}
\description{
Title
}
\examples{
\dontrun{
library(wcomp)

set.seed(1)
data = wcomp.generate_example_dataset(m1 = 100,
       n_X = 50,
       n_Y = 50,
       signal_strength_as_change_in_microbial_load = 0.1)



} 
}
