% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceScore_Plot.R
\name{wcomp.plot_reference_scores}
\alias{wcomp.plot_reference_scores}
\title{Title}
\usage{
wcomp.plot_reference_scores(ref_select, label, quantiles_to_plot = c(0.5,
  0.7, 0.9), breaks_param = 30)
}
\arguments{
\item{ref_select}{}

\item{label}{}

\item{quantiles_to_plot}{}

\item{breaks_param}{}
}
\description{
Title
}
