% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcomp_main_function.R
\name{wcomp.test}
\alias{wcomp.test}
\title{Title}
\usage{
wcomp.test(X, y, ind_reference_taxa, z = NULL, test = "Wilcoxon",
  q = 0.05, nr_perm = 1/(q/(ncol(X) - length(ind_reference_taxa))),
  disable_DSFDR = F, verbose = F)
}
\arguments{
\item{X}{}

\item{y}{}

\item{ind_reference_taxa}{}

\item{z}{}

\item{test}{}

\item{q}{}

\item{nr_perm}{}

\item{disable_DSFDR}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
\details{

}
\references{
foo
}
\author{
foo
}
