% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dacomp_check_reference_set.R
\name{dacomp.check_reference_set_is_valid.k_groups}
\alias{dacomp.check_reference_set_is_valid.k_groups}
\title{Title}
\usage{
dacomp.check_reference_set_is_valid.k_groups(X_ref, Y, nr.perm = 10^4,
  verbose = F)
}
\arguments{
\item{X_ref}{}

\item{Y}{}

\item{nr.perm}{}

\item{verbose}{}
}
\description{
Title
}
\examples{
#' #' \dontrun{

library(dacomp)

set.seed(1)

data = dacomp.generate_example_dataset(m1 = 100, 
       n_X = 50, n_Y = 50,
       signal_strength_as_change_in_microbial_load = 0.1)

#select references: (may take a minute)

result.selected.references = dacomp.select_references(X = data$counts,
                                 median_SD_threshold = 0.6, #APPLICATION SPECIFIC
                                 verbose = T)

length(result.selected.references$selected_references)

#plot the reference selection scores (can be used to better set the median SD threshold...)
dacomp.plot_reference_scores(result.selected.references)

result.ref.validity = dacomp.check_reference_set_is_valid.k_groups(
                        X_ref = data$counts[,result.selected.references$selected_references],
                        Y = data$group_labels,
                        nr.perm = 10000,
                        verbose = T)
result.ref.validity

} 
}
