% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dacomp_select_references.R
\name{dacomp.select_references}
\alias{dacomp.select_references}
\title{Title}
\usage{
dacomp.select_references(X, median_SD_threshold, minimal_TA = 10,
  maximal_TA = 200, Pseudo_Count_used = 1, verbose = F,
  select_from = NULL, Previous_Reference_Selection_Object = NULL)
}
\arguments{
\item{X}{}

\item{median_SD_threshold}{}

\item{minimal_TA}{}

\item{maximal_TA}{}

\item{Pseudo_Count_used}{}

\item{verbose}{}

\item{select_from}{}
}
\value{
a list
\itemize{
\item{selected_references}{}
\item{mean_prevalence_over_the_sorted}{}
\item{min_abundance_over_the_sorted}{}
\item{ratio_matrix}{}
\item{scores}{}
\item{selected_MinAbundance}{}
\item{median_SD_threshold}{}
\item{minimal_TA}{}
\item{maximal_TA}{}
}
}
\description{
Title
}
\examples{
#' \dontrun{
library(dacomp)

set.seed(1)

data = dacomp.generate_example_dataset.two_sample(m1 = 100,
       n_X = 50,
       n_Y = 50,
       signal_strength_as_change_in_microbial_load = 0.1)

# Select references: (may take a minute)
result.selected.references = dacomp.select_references(X = data$counts,
                                                     median_SD_threshold = 0.6, #APPLICATION SPECIFIC
                                                     verbose = T)

length(result.selected.references$selected_references)

# Plot the reference selection scores (can also be used to better set the median SD threshold)
dacomp.plot_reference_scores(result.selected.references)

# Select a reference set with a different (lower) threshold. user can use the function argument
# 'Previous_Reference_Selection_Object' to provide a previous reference seleciton object for this data, to speed up computation.

result.selected.references.different.threshold = dacomp.select_references(X = data$counts,
          median_SD_threshold = 0.5, 
          verbose = F,
          Previous_Reference_Selection_Object = result.selected.references)


} 
}
