% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dacomp_main_function.R
\name{dacomp.test}
\alias{dacomp.test}
\title{Title}
\usage{
dacomp.test(X, y, ind_reference_taxa, test = DACOMP.TEST.NAME.WILCOXON,
  q = 0.05, nr_perm = 1/(q/(ncol(X) - length(ind_reference_taxa))),
  disable_DSFDR = F, user_defined_test_function = NULL, verbose = F)
}
\arguments{
\item{X}{}

\item{y}{}

\item{ind_reference_taxa}{}

\item{test}{}

\item{q}{}

\item{nr_perm}{}

\item{disable_DSFDR}{}

\item{verbose}{}
}
\description{
Title
}
\examples{
#' \dontrun{
library(dacomp)

set.seed(1)

data = dacomp.generate_example_dataset(m1 = 100,
       n_X = 50,
       n_Y = 50,
       signal_strength_as_change_in_microbial_load = 0.1)

#select references: (may take a minute)
result.selected.references = dacomp.select_references(X = data$counts,
                                                     median_SD_threshold = 0.6, #APPLICATION SPECIFIC
                                                     verbose = T)

length(result.selected.references$selected_references)

#plot the reference selection scores (can also be used to better set the median SD threshold)
dacomp.plot_reference_scores(result.selected.references)

#multiplicity correction levels for the BH and DS-FDR methods
q_BH = q_DSFDR = 0.1

#Perform testing:
result.test = dacomp.test(X = data$counts,
                      y = data$group_labels,
                      ind_reference_taxa = result.selected.references,
                      verbose = T,q = q_DSFDR)

rejected_BH = which(p.adjust(result.test$p.values.test,method = 'BH')<=q_BH)
rejected_DSFDR = result.test$dsfdr_rejected

# example with continous covariate
set.seed(1)

data = dacomp.generate_example_dataset_continuous(n = 100,m1 = 30,
signal_strength_as_change_in_microbial_load = 0.1)


result.selected.references = dacomp.select_references(X = data$counts,
                                                     median_SD_threshold = 0.6, #APPLICATION SPECIFIC
                                                     verbose = T)
#number of selected references
length(result.selected.references$selected_references)

#plot the reference selection scores (can also be used to better set the median SD threshold)
dacomp.plot_reference_scores(result.selected.references)

#multiplicity correction levels for the BH and DS-FDR methods
q_BH = q_DSFDR = 0.1

#Perform testing:
result.test = dacomp.test(X = data$counts,
                         y = data$covariate,test = DACOMP.TEST.NAME.SPEARMAN,
                         ind_reference_taxa = result.selected.references,
                         verbose = T,q = q_DSFDR)

rejected_BH = which(p.adjust(result.test$p.values.test,method = 'BH')<=q_BH)
rejected_DSFDR = result.test$dsfdr_rejected
}
}
