% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Functions.R
\name{print.dacomp.reference.selection.object}
\alias{print.dacomp.reference.selection.object}
\title{Print information on the selected reference set, given an object from type reference.selection.object}
\usage{
\method{print}{dacomp.reference.selection.object}(obj)
}
\arguments{
\item{obj}{Object returned by dacomp.select.references(...)}
}
\description{
Print information on the selected reference set, given an object from type reference.selection.object
}
