% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Functions.R
\name{print.dacomp.reference.validation.result.object}
\alias{print.dacomp.reference.validation.result.object}
\title{Print the results of an object from type reference.validation.result.object}
\usage{
\method{print}{dacomp.reference.validation.result.object}(obj)
}
\arguments{
\item{obj}{Object returned by dacomp.check_reference_set_is_valid.k_groups}
}
\value{

}
\description{
Print the results of an object from type reference.validation.result.object
}
