% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_references.R
\name{wcomp.select_references}
\alias{wcomp.select_references}
\title{Title}
\usage{
wcomp.select_references(X, median_SD_threshold = 1.3, minimal_TA = 10,
  maximal_TA = 200, Psuedo_Count_used = 1, verbose = F,
  select_from = NULL)
}
\arguments{
\item{X}{}

\item{median_SD_threshold}{}

\item{minimal_TA}{}

\item{maximal_TA}{}

\item{Psuedo_Count_used}{}

\item{verbose}{}

\item{select_from}{}
}
\description{
Title
}
